requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
google-cloud-bigquery
google-cloud
mlflow
s3fs
google.cloud

#Streamlit libraries
streamlit
folium
streamlit_folium
geocoder
bokeh
streamlit_bokeh_events
branca

# utilities
six>=1.14
joblib
memoized-property
termcolor
